+++ ruixing-dao/src/main/java/com/yintu/ruixing/dao/CustomerDutyDao.java
@@ -5,13 +5,13 @@ import com.yintu.ruixing.entity.CustomerDutyEntity;
 import java.util.List;
 
 public interface CustomerDutyDao {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(CustomerDutyEntity record);
 
     int insertSelective(CustomerDutyEntity record);
 
-    CustomerDutyEntity selectByPrimaryKey(Integer id);
+    CustomerDutyEntity selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(CustomerDutyEntity record);
